requirements.txt
langchain
langchain-core
langchain_openai
pytube
python-dotenv
openai